Move PeriodicTrigger into utils
